Add files via upload
break and continue statements are generally used in loops . Break is used if we have to exit the statement or the loop IMMEDIATELY.
continue is used to skip only the particular statement. if used in the loop it will skip the statement then will go the the next counter until the expression becomes false. while loop with continue statements declare counter two times.