operators_of_pointers
POINTER is the variable that is used to stored the address of the other variable. It is declared with "*" .Ex- int *ageptr;
only for declaration purpose , "*"  is used for pointers but while using it it will b used as normal variable without the "*". to store the address , should be using the "&" infront of the variable. ex- ageptr= &age;
the speciality of pointer is that you can also GET the value stored at the address or the memory location. the syntax would be just adding the '*' in front of variable.ex- *ageptr; will give the the value stored in the variable age.
Dont get CONFUSED in the declaration .for declaration purpose also , we will use the '*' but while using it in the program , it has differnt PURPOSE.