passing Structure as Argument to function .
program is ip . Also known as pass by address . Simple program with two fn taking structure as argument , one has normal variable and other has pointer variable as an argument. Since , pointer is of same name as that of other variable , no need to assign the address to it.
changes in fn 1 wont make any changes to the main() but in fn 2 it will change the original values.
because when you call By-Value it makes a copy and sends that into the function. You are now making changes to the copy not the original back in the Main(). A pointer, points directly to the memory address of the original﻿.