Structure_dec_def
Structure is the collection of different  elements of differnt variables. unlike array which is the collection of similar elements of same data type. iwe define the variable of the structure in the main fn (or any other fn)  and then through that variable we can access and initialize the variables of structure . syntax-struct student{ }  , ex- student stud ; stud.name= "john"; 
in other words , we can say that structure is the modified version of array since all the values will  be stored in the name of SAME variable name but they can  be of differnt datatype . It is a USER DEFINED datatype. ex- student stud;  ,  datatype of stud is student .