Nested_Structures_by_arrow_operator
if we have two structures and the 2 has 1 in it ,being defined (forming nested structures), then we can access both the structure variables through their pointers . look for the example. ex-rohanptr->addr.street; where rohanptr is the pointer variable of the 2 structure , addr is the variable of 1 struct.    
ex-rohanptr->addrptr->house_no; it has rohanptr as pointer variable of 2 struct. and addrptr as pointer variable of struct. 1 which has stored the addrss of ANOTHER variable of the SAME structure i.e. struct. 1.